防止可能的内存引用问题；
修正：多个命令同时到达时未能处理后续指令的问题；优化：对 Socket 写入进行串行化保护；

重构:多线程响应模式,利用多线程计算单线程IO响应; 修正:线程数量由于单字节类型识别错误; 新增: PING 指令正确响应;

PING 指令应返回 PONG ；

允许无参指令；带参 PING 支持；

补充可执行文件 CentOS 7.6 环境；

补充若干词汇；

Update README.md

Update README.md

删除二进制文件;

+++ src/config.h
@@ -4,7 +4,7 @@
 class config {
 public:
 	std::filesystem::path            droot;
-	std::uint8_t                     wrkct;
+	std::uint16_t                    wrkct;
 	boost::asio::ip::tcp::endpoint address;
 	std::filesystem::path            qqwry;
 	std::string                      jdict;
